refactor(util): consolidate util from 16 to 10 source files, no import-path change
internal/util is imported by 57 files across 23 packages (util.SafeGo alone has 37
call sites), so changing its import path would have a wide blast radius and collide
with four concurrent restructures. This reduces the folder count without moving the
package: the widely-used entry points stay exactly where callers expect them.

Three consolidations, all driven by the reference graph rather than filenames:

- binary{,_bson,_cbor,_messagepack,_protobuf}.go -> binary.go (5 -> 1). The four
  per-format detectors were minted by a single splitting commit (14386082,
  "refactor: split binary format detectors by protocol"), so the shared filename
  prefix is not independent evidence of a real seam. The call graph says they are
  one closed unit: DetectBinaryFormat is the sole entry point and dispatches to all
  four, and the binaryMarker table type crosses the messagepack/cbor file boundary.
  A subpackage was rejected because DetectBinaryFormat's only callers live in
  internal/capture, which is being restructured concurrently.
- duration.go -> time.go (2 -> 1). Timestamp parsing and duration formatting are
  one concern.
- maps.go -> strings.go (2 -> 1). Two dependency-free pure helpers over the same
  string/key vocabulary.

Also corrects two stale symbol names in the package doc, found while rewriting it:
doc.go advertised IsBinaryPayload and WriteJSONResponse, neither of which exists;
the real names are DetectBinaryFormat and JSONResponse. Documentation-only.

Behavior preservation is mechanical: the exported symbol set is byte-identical
before and after (14 funcs + 1 type), and the multiset of non-comment code lines is
identical at 353 lines. Package coverage is unchanged at 76.0% with no function at
0.0%. GOOS=linux/windows/darwin all build, covering the proc_unix/proc_windows pair.
