--- internal/util/duration.go
@@ -1,32 +1,0 @@
-// duration.go — Human-readable duration formatting utility.
-
-package util
-
-import (
-	"fmt"
-	"time"
-)
-
-// FormatDuration formats a duration as human-readable (e.g., "5s", "2m30s", "1h15m").
-func FormatDuration(d time.Duration) string {
-	if d < time.Second {
-		return fmt.Sprintf("%.1fs", d.Seconds())
-	}
-	if d < time.Minute {
-		return fmt.Sprintf("%ds", int(d.Seconds()))
-	}
-	if d < time.Hour {
-		mins := int(d.Minutes())
-		secs := int(d.Seconds()) % 60
-		if secs == 0 {
-			return fmt.Sprintf("%dm", mins)
-		}
-		return fmt.Sprintf("%dm%02ds", mins, secs)
-	}
-	hours := int(d.Hours())
-	mins := int(d.Minutes()) % 60
-	if mins == 0 {
-		return fmt.Sprintf("%dh", hours)
-	}
-	return fmt.Sprintf("%dh%02dm", hours, mins)
-}
